pyproject.toml
[project]
name = "tidb-strands-agent"
version = "0.1.0"
description = "Intelligent Sales Assistant to get sale and conversations data"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.45.1",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.12.13",
    "strands-agents[openai]>=1.0.0",
    "strands-agents>=1.0.0",
    "strands-agents-tools>=0.2.0",
    "nest-asyncio>=1.6.0",
    "pyyaml>=6.0.2",
    "pytidb>=0.0.9",
    "litellm>=1.74.9.post1",
]
